# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Artists Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for current artist
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PROPOSED, PAYMENT_PENDING, PAID, IN_PROGRESS, DELIVERED, CANCELLED]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"

    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # PAYMENTS
  /payments/checkout:
    post:
      operationId: createCheckout
      tags: [payments]
      summary: Create payment checkout
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCheckoutBody"
      responses:
        "200":
          description: Checkout created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResponse"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /payments/confirm:
    post:
      operationId: confirmPayment
      tags: [payments]
      summary: Confirm payment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmPaymentBody"
      responses:
        "200":
          description: Payment confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/order/{orderId}:
    get:
      operationId: getPaymentByOrder
      tags: [payments]
      summary: Get payment for an order
      security:
        - bearerAuth: []
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # MEDIA
...
  /reviews:
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review for a delivered order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /reviews/artist/{artistId}:
    get:
      operationId: getArtistReviews
      tags: [reviews]
      summary: Get all reviews for an artist
      parameters:
        - name: artistId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewListResponse"

  # DASHBOARD (wow endpoints)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    OrderStatus:
      type: string
      enum:
        - PROPOSED
        - PAYMENT_PENDING
        - PAID
        - IN_PROGRESS
        - DELIVERED
        - CANCELLED
    Order:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        artistId:
          type: string
        title:
          type: string
        description:
          type: string
        occasion:
          type: string
        names:
          type: array
          items:
            type: string
        referenceLinks:
          type: array
          items:
            type: string
        deadline:
          type: string
          format: date-time
        additionalInstructions:
          type: string
        basePrice:
          type: number
        status:
          $ref: "#/components/schemas/OrderStatus"
        clientName:
          type: string
        clientEmail:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - clientId
        - artistId
        - title
        - deadline
        - basePrice
        - status
        - createdAt
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
      required:
        - orders
        - total
    CreateOrderBody:
      type: object
      properties:
        artistId:
          type: string
        clientName:
          type: string
        clientEmail:
          type: string
          format: email
        title:
          type: string
        description:
          type: string
        occasion:
          type: string
        names:
          type: array
          items:
            type: string
        referenceLinks:
          type: array
          items:
            type: string
        deadline:
          type: string
          format: date-time
        additionalInstructions:
          type: string
        basePrice:
          type: number
      required:
        - artistId
        - clientName
        - clientEmail
        - title
        - deadline
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/OrderStatus"
        deliveryVideoUrl:
          type: string
      required:
        - status

    # PAYMENT SCHEMAS
    Payment:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        amount:
          type: integer
        currency:
          type: string
        status:
          type: string
          enum: [PENDING, CONFIRMED, REFUNDED, FAILED]
        provider:
          type: string
          enum: [stripe, asaas]
        transactionId:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orderId
        - amount
        - currency
        - status
        - provider
        - createdAt
    CreateCheckoutBody:
      type: object
      properties:
        orderId:
          type: string
        provider:
          type: string
          enum: [stripe, asaas]
      required:
        - orderId
    CheckoutResponse:
      type: object
      properties:
        checkoutUrl:
          type: string
        sessionId:
          type: string
        expiresAt:
          type: string
          format: date-time
      required:
        - checkoutUrl
        - sessionId
    ConfirmPaymentBody:
      type: object
      properties:
        sessionId:
          type: string
        provider:
          type: string
          enum: [stripe, asaas]
      required:
        - sessionId
        - provider

    # MEDIA SCHEMAS
...
    Review:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        artistId:
          type: string
        clientId:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - orderId
        - artistId
        - clientId
        - rating
        - createdAt
    ReviewListResponse:
      type: object
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        total:
          type: integer
        averageRating:
          type: number
      required:
        - reviews
        - total
        - averageRating
    CreateReviewBody:
      type: object
      properties:
        orderId:
          type: string
        clientId:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          maxLength: 500
      required:
        - orderId
        - clientId
        - rating

    # DASHBOARD SCHEMAS
...