# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Artists Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateOrderBody {
  artistId: string;
  clientName: string;
  clientEmail: string;
  title: string;
  description?: string;
  occasion?: string;
  names?: string[];
  referenceLinks?: string[];
  deadline: Date;
  additionalInstructions?: string;
  basePrice?: number;
}