# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Artists Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: string;
  clientId: string;
  artistId: string;
  title: string;
  description?: string;
  occasion?: string;
  names?: string[];
  referenceLinks?: string[];
  deadline: Date;
  additionalInstructions?: string;
  basePrice: number;
  status: OrderStatus;
  clientName?: string;
  clientEmail?: string;
  createdAt: Date;
  updatedAt?: Date;
}