# lib/api-zod/src/generated/types/updateOrderStatusBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Artists Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from "./orderStatus";

export interface UpdateOrderStatusBody {
  status: OrderStatus;
  deliveryVideoUrl?: string;
}